data_validation

+++ networksecurity/utils/main_utils/utils.py
@@ -14,13 +14,30 @@ def read_yaml_file(file_path:str)-> dict:
     except Exception as e:
         raise NetworkSecurityException(e,sys) from e
 
-def write_yaml_file(file_path:str,content:object,replace:bool=False)->None:
+# def write_yaml_file(file_path:str,content:object,replace:bool=False)->None:
+#     try:
+#         dir_path = os.path.dirname(file_path)
+#         os.makedirs(dir_path, exist_ok=True)
+#         if replace:
+#             if os.path.exists(file_path):
+#                 os.remove(file_path)
+#         # os.makedirs(file_path,exist_ok=True)
+#         with open(os.path.dirname(file_path), "w") as file:
+#             yaml.dump(content,file)
+#     except Exception as e:
+#         raise NetworkSecurityException(e,sys)
+
+def write_yaml_file(file_path: str, content: object, replace: bool = False) -> None:
     try:
-        if replace:
-            if os.path.exists(file_path):
-                os.remove(file_path)
-        os.makedirs(file_path,exist_ok=True)
-        with open(os.path.dirname(file_path), "w") as file:
-            yaml.dump(content,file)
+        dir_path = os.path.dirname(file_path)
+
+        os.makedirs(dir_path, exist_ok=True)
+
+        if replace and os.path.exists(file_path):
+            os.remove(file_path)
+
+        with open(file_path, "w") as file:
+            yaml.dump(content, file)
+
     except Exception as e:
-        raise NetworkSecurityException(e,sys)
+        raise NetworkSecurityException(e, sys)
